+++ Computer/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 ## [Unreleased]
 
+## [0.7.1] - 2026-09-15
+
 ### Changed
 
 - Use a plus icon for the toolbar's Create menu.
-
-## [0.7.1] - 2026-09-15
 
 ### Fixed